0.1325

+++ Assets/Scripts/SaveAndLoad/WorldManager.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Runtime.Serialization.Formatters.Binary;
 using System.Text;
 using System.Threading.Tasks;
+using Unity.Collections;
 using Unity.Entities;
 using Unity.NetCode;
 using Unity.VisualScripting;
@@ -15,9 +16,9 @@ public static class WorldManager
 {
     public static bool WorldExist(string name)
     {
-        if (Directory.Exists(SaveSystem.savesPath))
+        if (Directory.Exists(SavePaths.savesPath))
         {
-            var files = Directory.GetDirectories(SaveSystem.savesPath);
+            var files = Directory.GetDirectories(SavePaths.savesPath);
 
             foreach (var item in files)
             {
@@ -28,35 +29,32 @@ public static class WorldManager
     }
     public static void RemoveWorld(string name)
     {
-        string file = Path.Combine(SaveSystem.savesPath, name);
+        if(string.IsNullOrEmpty(name)) return;
+        string file = SavePaths.GetWorldPath(name);
+        Debug.Log(file);
         if (Directory.Exists(file))
         {
             Directory.Delete(file, true);
         }
     }
-
     public static bool ChangeName(string oldName, string newName)
     {
-        // if(oldName == newName) return true;
-        // string oldPath = Path.Combine(SaveSystem.savesPath, oldName);
-        // string newPath = Path.Combine(SaveSystem.savesPath, newName);
-        // if (Directory.Exists(oldPath) && !Directory.Exists(newPath))
-        // {
-        //     BinaryFormatter formatter = new BinaryFormatter();
+        if(oldName == newName) return true;
+        string oldPath = Path.Combine(SavePaths.savesPath, oldName);
+        string newPath = Path.Combine(SavePaths.savesPath, newName);
 
-        //     FileStream fileStream = new FileStream(SaveSystem.GetHeaderPath(oldPath), FileMode.Open);
-        //     HeaderData headerData = formatter.Deserialize(fileStream) as HeaderData;
-        //     fileStream.Close();
-
-        //     headerData.worldName = newName;
-
-        //     fileStream = new FileStream(SaveSystem.GetHeaderPath(oldPath), FileMode.Create);
-        //     formatter.Serialize(fileStream, headerData);    
-        //     fileStream.Close();
-
-        //     Directory.Move(oldPath, newPath);
-        //     return true;
-        // }
+        if (Directory.Exists(oldPath) && !Directory.Exists(newPath))
+        {
+            if(SaveIOThread.TryLoadHeader(oldName,out HeaderSave header))
+            {
+                header.worldName = (FixedString128Bytes)newName;
+                if(SaveIOThread.SaveHeader(oldName,header))
+                {
+                    Directory.Move(oldPath, newPath);
+                    return true;
+                }
+            }
+        }
         return false;
     }
 }
